Catch CancelledKeyException on JVM
After b7297fc, it’s possible that the write callback closes the socket
and afterwards we get an error when we try to determine if the socket is
still readable. (Which is a bit weird – shouldn’t it return false?) This
fixes that problem.

(I assume the opposite problem existed before, if you closed the socket
from the read callback, but the test program just didn’t do that.)