Support writing byte buffers in record-based instance
The write callback now accepts a string or a byte buffer; in the latter
case, the argument is forwarded unchanged to the underlying write
callback. This allows us to write the HTTP server in a way that doesn’t
send a trailing \r.

Observer how the other test cases don’t need to be updated: you can
define your instance function as accepting just a String write callback,
and because of how Ceylon’s function types make sense, that *just
works*. (For the same reason, I have deliberately not updated the
example in the startRecordBased documentation.)

This also fixes a tiny bug where the write callback didn’t pass the
error strategy to encodeBuffer.