Add test for packetBased
The Ceylon side implements a server that can respond to several kinds of
requests at once: the classic greeter we already have several variants
of, a simple cat mirror, and you can also close the connection or even
kill the server.

The C code isn’t pretty. It doesn’t have to be. It works. Pull requests
to improve it are of course welcome, but I myself won’t bother with it.