Add stub client
This will eventually become a client for the daemonizeProgram module.
Right now, it just bails if you’re not in UTF-8. (I don’t intend to add
non-UTF-8 support at all, but recognizing the case is a good start.)

(As an aside, I find it slightly amusing that this requires more include
lines than lines of code in the actual main function.)