Add stdin support to client
If STDIN is defined at compile time, the client will read from standard
input until end of file, and send all received input to the daemon.

This mechanism is of course fairly limited. If you need to decide
whether to read from standard input based on some other condition
(e. g., the presence or absence of an argument or environment variable),
or if you want to print something some message to the user before
launching the daemon (“enter X”), you’ll need to adapt the code. But
this preprocessor-based mechanism should at least suffice for the
simplest of cases.