Fix bug with packets split up over transmissions
We were resetting the length buffer too early, so if a packet was split
across multiple transmissions (= multiple readPacket calls), part of the
packet content would end up in the length buffer, resulting in a
permanent offset of the packet boundary (the protocol isn’t
self-synchronizing). The fix is just two lines of diff (move the clear()
to a different line), but I also added an assertion to sanity check the
fix and some extra output to aid debugging.