Initial version of languageServerProtocol module
For #1. Implemented yesterday on, where else, a train ride. It already
no longer matches the design I actually want, but I haven’t yet had the
time to put those musings into code, so I’ll add them to the issue and
commit this code as it is for now.

(No automatic tests, but the sendMessage script sends a message with the
right format (Content-Length), and I’ve manually tested that the module
can consume its own reply as request again.)